@@ -1,14 +1,13 @@
-﻿namespace Monoboy.Constants
+﻿namespace Monoboy.Constants;
+
+public static class Bit
 {
-    public static class Bit
-    {
-        public const byte Bit7 = 0b10000000;
-        public const byte Bit6 = 0b01000000;
-        public const byte Bit5 = 0b00100000;
-        public const byte Bit4 = 0b00010000;
-        public const byte Bit3 = 0b00001000;
-        public const byte Bit2 = 0b00000100;
-        public const byte Bit1 = 0b00000010;
-        public const byte Bit0 = 0b00000001;
-    }
+    public const byte Bit7 = 0b10000000;
+    public const byte Bit6 = 0b01000000;
+    public const byte Bit5 = 0b00100000;
+    public const byte Bit4 = 0b00010000;
+    public const byte Bit3 = 0b00001000;
+    public const byte Bit2 = 0b00000100;
+    public const byte Bit1 = 0b00000010;
+    public const byte Bit0 = 0b00000001;
 }
