@@ -1,10 +1,9 @@
-﻿namespace Monoboy.Constants
+﻿namespace Monoboy.Constants;
+
+public static class Flag
 {
-    public static class Flag
-    {
-        public const byte Z = Bit.Bit7;
-        public const byte N = Bit.Bit6;
-        public const byte H = Bit.Bit5;
-        public const byte C = Bit.Bit4;
-    }
+    public const byte Z = Bit.Bit7;
+    public const byte N = Bit.Bit6;
+    public const byte H = Bit.Bit5;
+    public const byte C = Bit.Bit4;
 }
