@@ -1,14 +1,13 @@
-﻿namespace Monoboy.Constants
+﻿namespace Monoboy.Constants;
+
+public static class LCDCBit
 {
-    public static class LCDCBit
-    {
-        public const byte LCDEnabled = Bit.Bit7;
-        public const byte WindowTilemap = Bit.Bit6;
-        public const byte WindowEnabled = Bit.Bit5;
-        public const byte Tileset = Bit.Bit4;
-        public const byte Tilemap = Bit.Bit3;
-        public const byte SpritesSize = Bit.Bit2;
-        public const byte SpritesEnabled = Bit.Bit1;
-        public const byte BackgroundEnabled = Bit.Bit0;
-    }
+    public const byte LCDEnabled = Bit.Bit7;
+    public const byte WindowTilemap = Bit.Bit6;
+    public const byte WindowEnabled = Bit.Bit5;
+    public const byte Tileset = Bit.Bit4;
+    public const byte Tilemap = Bit.Bit3;
+    public const byte SpritesSize = Bit.Bit2;
+    public const byte SpritesEnabled = Bit.Bit1;
+    public const byte BackgroundEnabled = Bit.Bit0;
 }
