@@ -1,10 +1,9 @@
-﻿namespace Monoboy.Constants
+﻿namespace Monoboy.Constants;
+
+public static class Mode
 {
-    public static class Mode
-    {
-        public const byte Hblank = 0;
-        public const byte Vblank = 1;
-        public const byte OAM = 2;
-        public const byte VRAM = 3;
-    }
+    public const byte Hblank = 0;
+    public const byte Vblank = 1;
+    public const byte OAM = 2;
+    public const byte VRAM = 3;
 }
