@@ -1,13 +1,12 @@
-﻿namespace Monoboy.Constants
+﻿namespace Monoboy.Constants;
+
+public static class StatBit
 {
-    public static class StatBit
-    {
-        public const byte CoincidenceInterrupt = Bit.Bit6;
-        public const byte OAMInterrupt = Bit.Bit5;
-        public const byte VBlankInterrupt = Bit.Bit4;
-        public const byte HBlankInterrupt = Bit.Bit3;
-        public const byte CoincidenceFlag = Bit.Bit2;
-        public const byte ModeFlag1 = Bit.Bit1;
-        public const byte ModeFlag0 = Bit.Bit0;
-    }
+    public const byte CoincidenceInterrupt = Bit.Bit6;
+    public const byte OAMInterrupt = Bit.Bit5;
+    public const byte VBlankInterrupt = Bit.Bit4;
+    public const byte HBlankInterrupt = Bit.Bit3;
+    public const byte CoincidenceFlag = Bit.Bit2;
+    public const byte ModeFlag1 = Bit.Bit1;
+    public const byte ModeFlag0 = Bit.Bit0;
 }
